+++ dmr/routing.py
@@ -1,7 +1,6 @@
-import dataclasses
 from collections.abc import Callable, Coroutine, Iterable, Sequence
 from http import HTTPStatus
-from typing import TYPE_CHECKING, Any, TypeAlias, TypeVar, cast, final, overload
+from typing import TYPE_CHECKING, Any, TypeAlias, TypeVar, cast, overload
 
 from django.http import HttpRequest, HttpResponse, HttpResponseBase
 from django.urls import include
@@ -13,6 +12,7 @@ from typing_extensions import override
 
 from dmr.errors import ErrorType, format_error
 from dmr.exceptions import InternalServerError, NotAcceptableError
+from dmr.internal.routing import URLExternal as _URLExternal
 from dmr.openapi.collector import controller_mapping_collector
 from dmr.openapi.objects import PathItem, Paths
 from dmr.openapi.openapi import OpenAPI
@@ -41,68 +41,13 @@ _DjangoView: TypeAlias = Callable[
 _SerializerT = TypeVar('_SerializerT', bound='BaseSerializer')
 
 
-@final
-@dataclasses.dataclass(slots=True, frozen=True)
-class URLExternal:
-    """
-    Represents an external URL that was added to the routing of DMR.
-
-    Prefer :func:`external_path` over using this class directly.
-    See :ref:`external-views` for more info.
-
-    .. versionadded:: 0.13.0
-    """
-
-    url: URLPattern
-    openapi: PathItem | None = dataclasses.field(kw_only=True)
-
-    def get_url_with_metadata(self) -> URLPattern:
-        """Get the url pattern with attached OpenAPI metadata."""
-        self.url.callback.__dmr_external_openapi__ = self.openapi  # type: ignore[attr-defined]
-        return self.url
-
-
-def external_path(
-    route: '_StrOrPromise',
-    view: _DjangoView,
-    *,
-    openapi: PathItem | None,
-    kwargs: dict[str, Any] | None = None,
-    name: str | None = None,
-) -> URLExternal:
-    """
-    Add an external path onto the DMR routing system.
-
-    Parameters:
-        route: String route for the view.
-        view: Function or class view, supports both sync and async callables.
-        openapi: OpenAPI metadata to show in the spec.
-            Or ``None`` to hide this endpoint.
-        kwargs: Init kwargs for the view.
-        name: Name to resolve this URL.
-
-    Prefer this function over using :class:`URLExternal` directly.
-    See :ref:`external-views` for more info.
-
-    .. versionadded:: 0.13.0
-    """
-    return URLExternal(
-        _django_path(  # pyrefly: ignore[no-matching-overload]
-            route,
-            view,  # type: ignore[arg-type]
-            kwargs=kwargs,
-            name=name,
-        ),
-        openapi=openapi,
-    )
-
-
 class Router:
     """
     Collection of HTTP routes for REST framework.
 
     Attributes:
         prefix: URL prefix for all routes (e.g., 'api/v1/').
+            Defaults to empty string ``''``.
         urls: Sequence of URL patterns and resolvers.
         tags: Optional sequence of tags to group operations in OpenAPI.
             These are merged with endpoint-level tags.
@@ -118,7 +63,7 @@ class Router:
         Added *tags* and *deprecated* parameters.
 
     .. versionchanged:: 0.13.0
-        Now you can pass :class:`URLExternal` objects in *urls*.
+        Now you can pass :func:`external_path` objects in *urls*.
         Also accept any :class:`collections.abc.Sequence` as *tags*.
         *urls* parameter is now optional.
 
@@ -128,8 +73,8 @@ class Router:
 
     def __init__(
         self,
-        prefix: str,
-        urls: Iterable[_AnyPattern | URLExternal] = (),
+        prefix: str = '',
+        urls: Iterable[_AnyPattern | _URLExternal] = (),
         *,
         tags: Sequence[str] | None = None,
         deprecated: bool = False,
@@ -191,23 +136,82 @@ class Router:
 
         .. versionadded:: 0.13.0
         """
+        self.urls.append(
+            router.to_urlpatterns(namespace=namespace, app_name=app_name),
+        )
+
+    def to_urlpatterns(
+        self,
+        *,
+        namespace: str | None = None,
+        app_name: str | None = None,
+    ) -> URLResolver:
+        """
+        Convert router instance into ``urlpatterns`` include API.
+
+        Can be used to include one router into another.
+        Or to include a router into the final ``urlpatterns`` list.
+
+        Automatically uses our own faster :func:`path` function.
+
+        .. versionadded:: 0.14.0
+        """
         if app_name is None and namespace is not None:
             app_name = namespace
 
-        arg = router.urls if app_name is None else (router.urls, app_name)
-        self.urls.append(path(router.prefix, include(arg, namespace=namespace)))
+        path_spec = self.urls if app_name is None else (self.urls, app_name)
+        return path(self.prefix, include(path_spec, namespace=namespace))
 
     def _maybe_process_external(
         self,
-        urls: Iterable[_AnyPattern | URLExternal],
+        urls: Iterable[_AnyPattern | _URLExternal],
     ) -> list[_AnyPattern]:
         django_like_urls: list[_AnyPattern] = []
         for url in urls:
-            if isinstance(url, URLExternal):
+            if isinstance(url, _URLExternal):
                 django_like_urls.append(url.get_url_with_metadata())
             else:
                 django_like_urls.append(url)
         return django_like_urls
+
+
+def external_path(
+    route: '_StrOrPromise',
+    view: _DjangoView,
+    *,
+    openapi: PathItem | None,
+    kwargs: dict[str, Any] | None = None,
+    name: str | None = None,
+) -> '_URLExternal':
+    """
+    Add an external path onto the DMR routing system.
+
+    Automatically uses our own faster :func:`path` function.
+
+    Parameters:
+        route: String route for the view.
+        view: Function or class view, supports both sync and async callables.
+        openapi: OpenAPI metadata to show in the spec.
+            Or ``None`` to hide this endpoint.
+        kwargs: Init kwargs for the view.
+        name: Name to resolve this URL.
+
+    .. important::
+
+        This function only works when including
+        a URL into our own :class:`Router` objects,
+        not into the Django own ``urlpatterns``.
+
+        Django check ``urls.E004`` covers this statically.
+
+    See :ref:`external-views` for more info.
+
+    .. versionadded:: 0.13.0
+    """
+    return _URLExternal(
+        path(route, view, kwargs=kwargs, name=name),
+        openapi=openapi,
+    )
 
 
 # We mimic django's name here: