+++ docs/examples/external_views/classes.py
@@ -1,7 +1,6 @@
 import random
 
 from django.http import HttpRequest, JsonResponse
-from django.urls import include
 from django.views import View
 
 from dmr.openapi import build_schema, load_schema
@@ -35,7 +34,7 @@ schema = build_schema(router)
 
 urlpatterns = [
     # Register our router in the final url patterns:
-    path(router.prefix, include((router.urls, 'test_app'), namespace='api')),
+    router.to_urlpatterns(namespace='api'),
     # Add swagger:
     path('docs/openapi.json/', OpenAPIJsonView.as_view(schema), name='openapi'),
 ]